[feature] Temporary - crop yield map display in a conditional and automate crop yield display as last simulation year
The input of the function used to display results in the main and example scripts is not called identically for the different crop models, so I put a conditional to avoid having to manually adapt the scripts every time we switch crop models.

Also automatically select the last simulation year to display the crop yield, based on the parameters dictionary. Until now, it was hard coded. It made it so the user could run into an error when the hard-coded year did not match the time range of the simulation. Advanced users can still edit the date of the yield map display as needed.