[feature] New module sky_model with Reinhart sky discretization scheme
* Reinhart sky discretization scheme is in a new module sky_model in a new class ReinhartSky. 
Its only input is multiplying factor MF, that is used to multiply the 145 patches from the "original" Tregenza sky subdivision scheme into as many patches (e.g. MF = 4 subdivides each of the patches of the Tregenza scheme in 4). 
Only the top patch is not subdivided like the others, but its size decreases as MF increases.

* Conversion function from spherical to cartesian coordinates is in new file conversion_functions.py.
Inputs are documented in the function docstring.
Conversion functions should be stored there and reused as much as possible to reduce risks of errors.


Closes #116