[fix] Adding the new observed data folder with the example csv
Add a new folder in the benchmarking one where the csv with the observed reference data used for comparison (computing RMSE...) should be place. It had not been placed here because csv were hidden by the gitignore. Now the branch is up to date and should be ready to use and merge.