[fix] Error in number of Y panels generated when fixed to 8.
When NumbeYofpanels was set to 8, 9 panels were generated because of a bug in MODULES.PHOTOVOLTAICS.configuration.py.creat_block_of_panels. The formula used to space the panels evenly around 0 generated rounding errors (float 32) that could lead to the generation of an additional panel (case numberY=8). Slight modification of the formulas and no more float32). Now generates 8 panels correctly.