[ci] Improve CI jobs runtime from ca 20 min to ca 3 min by fixing caching strategy
Changed caching parameters and strategy.
The cache is stored under /cache in the server environment and shared between runners.
The former configuration took lots of time to save packages and virtual env to cache, but in the wrong location and it was deleted so never used.
The cache will be overwritten in case environment_unix.yml's contents change.