[feature] Adding a function to update apex_grazed.
Adds a function to update apex_grazed to 1 if cut during reproductive growth and apex above cut_height. This function is defined un Plant.py but is called in Management.py befor applying cut to verify that the height of reprodcutive part is above cut_height. NB : when using the frequency example, no apex is cut. Also, when using panels, heterogeneity because of shade, biomass differ so conditions for apex_grazed too.