[fix] Enforce Type field in input files
Raise an error if the "type" field is misspelled. Before, in case of misspelling, the parameter was not read but there was no error, which made for hard debugging.

The error states in which parameter there is a spelling error and in which file it is, as well as all the accepted options.

Closes #180