[perf] Reduction the number of multi_ray_trace
The intercepted rays retain information about the type of object intercepted (PV, diffuser, etc.),
which reduces the number of calls to the multi_ray_trace method. One multi_ray_trace per geometry
and per direction type (1 for the direction of the sun and 1 for the directions of the discrete sky).